[Core] Accept arbitrary Pylint args
Supporting only a restricted set of Pylint command-line arguments means
less customisation is available to consumers, in return for a small
saving on boilerplate.

In particular, new features added to Pylint may be available on some
consumers and not others (i.e. running different Pylint versions),
meaning we have to support a particular subset of command-line arguments
compatible with a specific version range.

A specific example is the `--recursive` flag that is available as of
Pylint v2.13.5.

So eschew a fine-grained subset of command-line arguments and instead
allow consumers to pass whatever arguments to Pylint they like.

Hence add `shell-quote` as a new dependency, for use in properly
splitting the arguments string for use in a subprocess call.

Signed-off-by: David Feltell <david.feltell@foundry.com>